Re-organize examples
Add more narrative content. Accept that we only got a single example for this version of the IG.

List examples manually (there are not that many of them, we can do this). This also helps populate the local TOC of the page.

Place the dragon warning to the end of page. Not sure this is optimal, but on the top of the page it collided with the page TOC box.