Strip off id from example capability statement
This is only added for the IG publisher to work, and not part of the capability statement as published by Apotti / Epic.

The publisher does add metadata like publisher, contact, and jurisdiction into the resource, though. I think we can live with this, and link to the original "live" version of the Apotti metadata, for anyone who wants to see how it really looks like.